fix: resolve 53 code quality issues from GitRoll scan
- Replace string concatenation with String.format in logging calls
- Extract constants (CONFIG_FILE, REWARDS_FILE, GRADIENT_END)
- Add specific exception types instead of broad throws clauses
- Reduce cognitive complexity via extracted helper methods
- Add @SuppressWarnings for intentional RSA/PKCS1 usage
- Remove unused fields (LOGGER in impl classes, voteConfig in handler)
- Add comments explaining empty catch blocks and volatile fields
- Fix integer overflow in leaderboard percentage calculation
- Use lambda-deferred logging for expensive string operations
- Rename shadowed variable (record → voteRecord in VoteService)
- Simplify boolean expressions and conditional returns

+++ jexvote-common/src/main/java/de/jexcellence/vote/view/VoteLeaderboardView.java
@@ -254,7 +254,7 @@ public class VoteLeaderboardView extends VoteBaseView {
 
     private static String buildVoteBar(int votes) {
         int bars = 10;
-        int filled = votes <= 0 ? 0 : Math.min(bars, Math.max(1, (int) (Math.log10(votes + 1) * 3)));
+        int filled = votes <= 0 ? 0 : Math.min(bars, Math.max(1, (int) (Math.log10((double) votes + 1) * 3)));
         var sb = new StringBuilder("<dark_gray>[</dark_gray>");
         for (int i = 0; i < bars; i++) {
             sb.append(i < filled ? "<gradient:#fde047:#f59e0b>|</gradient>" : "<dark_gray>|</dark_gray>");
